refactor: api 일부 복수형 변경
